WIP: Fix compiler warnings

+++ include/utils.h
@@ -30,8 +30,8 @@
 #define NAPI_URL                   "https://napi.nbg01.v10lator.de/v2/"
 #define NUSSPLI_COPYRIGHT          "© 2020-2023 V10lator <v10lator@myway.de>"
 
-#define CUSTOM_MCP_ERROR_EOM       0xDEAD0001
-#define CUSTOM_MCP_ERROR_CANCELLED 0xDEAD0002
+#define CUSTOM_MCP_ERROR_EOM       ((int)0xDEAD0001)
+#define CUSTOM_MCP_ERROR_CANCELLED ((int)0xDEAD0002)
 
 #ifdef NUSSPLI_DEBUG
 #include <whb/log.h>
